fix OSX build error

+++ db/flush_job.cc
@@ -185,7 +185,8 @@ void FlushJob::PickMemTable() {
   edit_->SetColumnFamily(cfd_->GetID());
 
   // path 0 for level 0 file.
-  meta_.fd = FileDescriptor(versions_->NewFileNumber(), path_id_, 0);
+  meta_.fd = FileDescriptor(versions_->NewFileNumber(), 
+      static_cast<uint32_t>(path_id_), 0);
 
   base_ = cfd_->current();
   base_->Ref();  // it is likely that we do not need this reference
